src/send-grid/requirements.txt
functions-framework==3.*
google-cloud-logging==3.10.*
sendgrid==6.*